--- tests/test_console_encoding_factory.py
@@ -1,318 +1,0 @@
-"""Console Encoding and Factory Deep Coverage Tests
-
-Console系UIモジュールの深度カバレッジ向上。
-target: console_encoding 28%→60%+, console_factory 59%→80%+, console_interaction 50%→75%+
-"""
-
-import sys
-from io import StringIO
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-
-class TestConsoleEncodingDeep:
-    """ConsoleEncoding深度テスト - 28%→60%+目標"""
-
-    def test_console_encoding_initialization(self):
-        """ConsoleEncoding初期化テスト"""
-        from kumihan_formatter.ui.console_encoding import ConsoleEncoding
-
-        encoding = ConsoleEncoding()
-        assert encoding is not None
-
-    def test_console_encoding_detection(self):
-        """エンコーディング検出テスト"""
-        from kumihan_formatter.ui.console_encoding import ConsoleEncoding
-
-        encoding = ConsoleEncoding()
-
-        # エンコーディング検出メソッドテスト
-        detection_methods = [
-            "detect_encoding",
-            "get_encoding",
-            "detect_file_encoding",
-            "auto_detect",
-            "guess_encoding",
-        ]
-
-        test_texts = [
-            b"Hello World",  # ASCII
-            "Hello World".encode("utf-8"),  # UTF-8
-            "日本語テキスト".encode("utf-8"),  # Japanese UTF-8
-        ]
-
-        for method_name in detection_methods:
-            if hasattr(encoding, method_name):
-                method = getattr(encoding, method_name)
-                if callable(method):
-                    for test_text in test_texts:
-                        try:
-                            result = method(test_text)
-                            if result is not None:
-                                assert isinstance(result, str)
-                        except Exception:
-                            pass
-
-    def test_console_encoding_conversion(self):
-        """エンコーディング変換テスト"""
-        from kumihan_formatter.ui.console_encoding import ConsoleEncoding
-
-        encoding = ConsoleEncoding()
-
-        # エンコーディング変換メソッドテスト
-        conversion_methods = [
-            "convert_encoding",
-            "encode_text",
-            "decode_text",
-            "to_utf8",
-            "from_utf8",
-            "safe_decode",
-        ]
-
-        test_content = "テスト文字列"
-
-        for method_name in conversion_methods:
-            if hasattr(encoding, method_name):
-                method = getattr(encoding, method_name)
-                if callable(method):
-                    try:
-                        if "decode" in method_name:
-                            result = method(test_content.encode("utf-8"))
-                        else:
-                            result = method(test_content)
-
-                        if result is not None:
-                            assert isinstance(result, (str, bytes))
-                    except Exception:
-                        try:
-                            result = method(test_content, "utf-8")
-                        except:
-                            pass
-
-    def test_console_encoding_validation(self):
-        """エンコーディング検証テスト"""
-        from kumihan_formatter.ui.console_encoding import ConsoleEncoding
-
-        encoding = ConsoleEncoding()
-
-        # エンコーディング検証メソッドテスト
-        validation_methods = [
-            "validate_encoding",
-            "is_valid_encoding",
-            "check_encoding",
-            "supports_encoding",
-            "is_supported",
-        ]
-
-        test_encodings = ["utf-8", "ascii", "latin-1", "cp932", "invalid-encoding"]
-
-        for method_name in validation_methods:
-            if hasattr(encoding, method_name):
-                method = getattr(encoding, method_name)
-                if callable(method):
-                    for test_encoding in test_encodings:
-                        try:
-                            result = method(test_encoding)
-                            assert isinstance(result, bool)
-                        except Exception:
-                            pass
-
-
-class TestConsoleFactoryDeep:
-    """ConsoleFactory深度テスト - 59%→80%+目標"""
-
-    def test_console_factory_initialization(self):
-        """ConsoleFactory初期化テスト"""
-        from kumihan_formatter.ui.console_factory import ConsoleFactory
-
-        factory = ConsoleFactory()
-        assert factory is not None
-
-    def test_console_factory_creation_methods(self):
-        """Console作成メソッドテスト"""
-        from kumihan_formatter.ui.console_factory import ConsoleFactory
-
-        factory = ConsoleFactory()
-
-        # Console作成メソッドテスト
-        creation_methods = [
-            "create_console",
-            "get_console",
-            "make_console",
-            "create_ui",
-            "get_ui",
-            "build_console",
-        ]
-
-        for method_name in creation_methods:
-            if hasattr(factory, method_name):
-                method = getattr(factory, method_name)
-                if callable(method):
-                    try:
-                        result = method()
-                        assert result is not None
-                    except Exception:
-                        try:
-                            result = method("default")
-                            assert result is not None
-                        except:
-                            pass
-
-    def test_console_factory_configuration(self):
-        """Factory設定テスト"""
-        from kumihan_formatter.ui.console_factory import ConsoleFactory
-
-        factory = ConsoleFactory()
-
-        # 設定メソッドテスト
-        config_methods = [
-            "configure",
-            "set_config",
-            "setup",
-            "initialize",
-            "set_options",
-            "apply_config",
-        ]
-
-        test_config = {"encoding": "utf-8", "color": True, "verbose": False}
-
-        for method_name in config_methods:
-            if hasattr(factory, method_name):
-                method = getattr(factory, method_name)
-                if callable(method):
-                    try:
-                        method(test_config)
-                    except Exception:
-                        try:
-                            method(**test_config)
-                        except:
-                            pass
-
-    def test_console_factory_types(self):
-        """異なるConsoleタイプ作成テスト"""
-        from kumihan_formatter.ui.console_factory import ConsoleFactory
-
-        factory = ConsoleFactory()
-
-        # 異なるタイプのConsole作成テスト
-        console_types = ["basic", "enhanced", "minimal", "full", "debug"]
-
-        for console_type in console_types:
-            try:
-                if hasattr(factory, "create_console"):
-                    result = factory.create_console(console_type)
-                elif hasattr(factory, "get_console"):
-                    result = factory.get_console(console_type)
-                else:
-                    # タイプ指定メソッドを探す
-                    type_method = f"create_{console_type}_console"
-                    if hasattr(factory, type_method):
-                        method = getattr(factory, type_method)
-                        result = method()
-                        assert result is not None
-            except Exception:
-                # タイプが存在しない場合は許容
-                pass
-
-
-class TestConsoleInteractionDeep:
-    """ConsoleInteraction深度テスト - 50%→75%+目標"""
-
-    def test_console_interaction_initialization(self):
-        """ConsoleInteraction初期化テスト"""
-        from kumihan_formatter.ui.console_interaction import ConsoleInteraction
-
-        interaction = ConsoleInteraction()
-        assert interaction is not None
-
-    def test_console_interaction_user_input(self):
-        """ユーザー入力処理テスト"""
-        from kumihan_formatter.ui.console_interaction import ConsoleInteraction
-
-        interaction = ConsoleInteraction()
-
-        # ユーザー入力メソッドテスト
-        input_methods = [
-            "get_user_input",
-            "read_input",
-            "prompt_user",
-            "ask_question",
-            "get_response",
-        ]
-
-        with patch("builtins.input", return_value="test response"):
-            for method_name in input_methods:
-                if hasattr(interaction, method_name):
-                    method = getattr(interaction, method_name)
-                    if callable(method):
-                        try:
-                            result = method("Question: ")
-                            assert isinstance(result, str)
-                        except Exception:
-                            try:
-                                result = method()
-                                assert isinstance(result, str)
-                            except:
-                                pass
-
-    def test_console_interaction_choices(self):
-        """選択肢処理テスト"""
-        from kumihan_formatter.ui.console_interaction import ConsoleInteraction
-
-        interaction = ConsoleInteraction()
-
-        # 選択肢メソッドテスト
-        choice_methods = [
-            "select_option",
-            "choose_from_list",
-            "get_choice",
-            "select_from_menu",
-            "pick_option",
-        ]
-
-        test_choices = ["Option 1", "Option 2", "Option 3"]
-
-        with patch("builtins.input", return_value="1"):
-            for method_name in choice_methods:
-                if hasattr(interaction, method_name):
-                    method = getattr(interaction, method_name)
-                    if callable(method):
-                        try:
-                            result = method(test_choices)
-                            assert result is not None
-                        except Exception:
-                            try:
-                                result = method("Select:", test_choices)
-                            except:
-                                pass
-
-    def test_console_interaction_confirmation(self):
-        """確認処理テスト"""
-        from kumihan_formatter.ui.console_interaction import ConsoleInteraction
-
-        interaction = ConsoleInteraction()
-
-        # 確認メソッドテスト
-        confirm_methods = [
-            "confirm_action",
-            "ask_confirmation",
-            "get_yes_no",
-            "confirm",
-            "verify_action",
-        ]
-
-        with patch("builtins.input", side_effect=["y", "yes", "n", "no"]):
-            for method_name in confirm_methods:
-                if hasattr(interaction, method_name):
-                    method = getattr(interaction, method_name)
-                    if callable(method):
-                        try:
-                            result = method("Confirm action?")
-                            assert isinstance(result, bool)
-                        except Exception:
-                            try:
-                                result = method()
-                                assert isinstance(result, bool)
-                            except:
-                                pass
