--- tests/test_console_ui_operations.py
@@ -1,429 +1,0 @@
-"""Console UI and Operations Deep Coverage Tests
-
-Console系UIモジュールの深度カバレッジ向上。
-target: console_ui 64%→80%+, console_operations 32%→60%+
-"""
-
-import sys
-from io import StringIO
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-
-class TestConsoleUIDeep:
-    """ConsoleUI深度テスト - 64%→80%+目標"""
-
-    def test_console_ui_initialization(self):
-        """ConsoleUI初期化テスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-        assert ui is not None
-
-        # 基本属性確認
-        assert hasattr(ui, "print") or hasattr(ui, "output")
-        assert hasattr(ui, "input") or hasattr(ui, "get_input")
-
-    def test_console_ui_print_methods(self):
-        """ConsoleUI出力メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # 標準出力をキャプチャ
-        with patch("sys.stdout", new_callable=StringIO) as mock_stdout:
-            # 基本print系メソッドテスト
-            print_methods = ["print", "output", "write", "display"]
-
-            for method_name in print_methods:
-                if hasattr(ui, method_name):
-                    method = getattr(ui, method_name)
-                    if callable(method):
-                        try:
-                            method("Test message")
-                        except Exception:
-                            # パラメータが違う場合の代替テスト
-                            try:
-                                method("Test", "message")
-                            except:
-                                pass
-
-    def test_console_ui_error_methods(self):
-        """ConsoleUIエラー出力メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # エラー出力メソッドテスト
-        error_methods = ["print_error", "error", "show_error", "display_error"]
-
-        with patch("sys.stderr", new_callable=StringIO):
-            for method_name in error_methods:
-                if hasattr(ui, method_name):
-                    method = getattr(ui, method_name)
-                    if callable(method):
-                        try:
-                            method("Test error message")
-                        except Exception:
-                            try:
-                                method("Error", "details")
-                            except:
-                                pass
-
-    def test_console_ui_warning_methods(self):
-        """ConsoleUI警告出力メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # 警告出力メソッドテスト
-        warning_methods = ["print_warning", "warning", "warn", "show_warning"]
-
-        for method_name in warning_methods:
-            if hasattr(ui, method_name):
-                method = getattr(ui, method_name)
-                if callable(method):
-                    try:
-                        method("Test warning message")
-                    except Exception:
-                        pass
-
-    def test_console_ui_success_methods(self):
-        """ConsoleUI成功出力メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # 成功出力メソッドテスト
-        success_methods = [
-            "print_success",
-            "success",
-            "show_success",
-            "display_success",
-        ]
-
-        for method_name in success_methods:
-            if hasattr(ui, method_name):
-                method = getattr(ui, method_name)
-                if callable(method):
-                    try:
-                        method("Test success message")
-                    except Exception:
-                        pass
-
-    def test_console_ui_input_methods(self):
-        """ConsoleUI入力メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # 入力メソッドテスト
-        input_methods = ["input", "get_input", "read_input", "prompt"]
-
-        # 標準入力をモック
-        with patch("builtins.input", return_value="test input"):
-            for method_name in input_methods:
-                if hasattr(ui, method_name):
-                    method = getattr(ui, method_name)
-                    if callable(method):
-                        try:
-                            result = method("Enter something: ")
-                            assert isinstance(result, str)
-                        except Exception:
-                            try:
-                                result = method()
-                                assert isinstance(result, str)
-                            except:
-                                pass
-
-    def test_console_ui_confirmation_methods(self):
-        """ConsoleUI確認メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # 確認メソッドテスト
-        confirm_methods = ["confirm", "ask_confirmation", "yes_no", "ask_yes_no"]
-
-        # Yes/No入力をモック
-        with patch("builtins.input", side_effect=["y", "n", "yes", "no"]):
-            for method_name in confirm_methods:
-                if hasattr(ui, method_name):
-                    method = getattr(ui, method_name)
-                    if callable(method):
-                        try:
-                            result = method("Continue?")
-                            assert isinstance(result, bool)
-                        except Exception:
-                            try:
-                                result = method()
-                                assert isinstance(result, bool)
-                            except:
-                                pass
-
-    def test_console_ui_progress_methods(self):
-        """ConsoleUIプログレス表示メソッドテスト"""
-        from kumihan_formatter.ui.console_ui import ConsoleUI
-
-        ui = ConsoleUI()
-
-        # プログレス表示メソッドテスト
-        progress_methods = [
-            "show_progress",
-            "progress",
-            "update_progress",
-            "display_progress",
-        ]
-
-        for method_name in progress_methods:
-            if hasattr(ui, method_name):
-                method = getattr(ui, method_name)
-                if callable(method):
-                    try:
-                        method(50, 100, "Processing...")
-                    except Exception:
-                        try:
-                            method(50)
-                        except:
-                            pass
-
-
-class TestConsoleOperationsDeep:
-    """ConsoleOperations深度テスト - 32%→60%+目標"""
-
-    def test_console_operations_initialization(self):
-        """ConsoleOperations初期化テスト"""
-        from kumihan_formatter.ui.console_operations import ConsoleOperations
-
-        ops = ConsoleOperations()
-        assert ops is not None
-
-    def test_console_operations_file_operations(self):
-        """ファイル操作メソッドテスト"""
-        from kumihan_formatter.ui.console_operations import ConsoleOperations
-
-        ops = ConsoleOperations()
-
-        # ファイル操作メソッドテスト
-        file_methods = [
-            "read_file",
-            "write_file",
-            "copy_file",
-            "delete_file",
-            "file_exists",
-            "create_directory",
-            "list_files",
-        ]
-
-        for method_name in file_methods:
-            if hasattr(ops, method_name):
-                method = getattr(ops, method_name)
-                if callable(method):
-                    try:
-                        # 安全なテストパスで実行
-                        if "read" in method_name:
-                            result = method("test.txt")
-                        elif "write" in method_name:
-                            method("test.txt", "content")
-                        elif "exists" in method_name:
-                            result = method("test.txt")
-                            assert isinstance(result, bool)
-                        elif "list" in method_name:
-                            result = method(".")
-                            assert isinstance(result, (list, tuple))
-                    except (FileNotFoundError, PermissionError):
-                        # ファイル操作エラーは期待される
-                        pass
-                    except Exception:
-                        # その他のエラーは許容
-                        pass
-
-    def test_console_operations_text_processing(self):
-        """テキスト処理メソッドテスト"""
-        from kumihan_formatter.ui.console_operations import ConsoleOperations
-
-        ops = ConsoleOperations()
-
-        # テキスト処理メソッドテスト
-        text_methods = [
-            "format_text",
-            "clean_text",
-            "validate_text",
-            "process_text",
-            "encode_text",
-            "decode_text",
-            "normalize_text",
-        ]
-
-        test_text = "Test text content"
-
-        for method_name in text_methods:
-            if hasattr(ops, method_name):
-                method = getattr(ops, method_name)
-                if callable(method):
-                    try:
-                        result = method(test_text)
-                        assert isinstance(result, (str, bool, dict))
-                    except Exception:
-                        # エラーは許容
-                        pass
-
-    def test_console_operations_conversion_methods(self):
-        """変換操作メソッドテスト"""
-        from kumihan_formatter.ui.console_operations import ConsoleOperations
-
-        ops = ConsoleOperations()
-
-        # 変換メソッドテスト
-        conversion_methods = [
-            "convert_to_html",
-            "convert_to_markdown",
-            "convert_format",
-            "apply_template",
-            "render_output",
-        ]
-
-        test_content = "# Test\n\nContent"
-
-        for method_name in conversion_methods:
-            if hasattr(ops, method_name):
-                method = getattr(ops, method_name)
-                if callable(method):
-                    try:
-                        result = method(test_content)
-                        assert isinstance(result, str)
-                    except Exception:
-                        try:
-                            result = method(test_content, "template")
-                        except:
-                            pass
-
-    def test_console_operations_validation_methods(self):
-        """検証操作メソッドテスト"""
-        from kumihan_formatter.ui.console_operations import ConsoleOperations
-
-        ops = ConsoleOperations()
-
-        # 検証メソッドテスト
-        validation_methods = [
-            "validate_syntax",
-            "check_syntax",
-            "validate_file",
-            "validate_input",
-            "validate_output",
-        ]
-
-        for method_name in validation_methods:
-            if hasattr(ops, method_name):
-                method = getattr(ops, method_name)
-                if callable(method):
-                    try:
-                        result = method("test content")
-                        assert isinstance(result, (bool, dict, list))
-                    except Exception:
-                        pass
-
-
-class TestConsoleMessagingDeep:
-    """ConsoleMessaging深度テスト - 43%→70%+目標"""
-
-    def test_console_messaging_initialization(self):
-        """ConsoleMessaging初期化テスト"""
-        from kumihan_formatter.ui.console_messaging import ConsoleMessaging
-
-        messaging = ConsoleMessaging()
-        assert messaging is not None
-
-    def test_console_messaging_message_types(self):
-        """メッセージタイプ別テスト"""
-        from kumihan_formatter.ui.console_messaging import ConsoleMessaging
-
-        messaging = ConsoleMessaging()
-
-        # メッセージタイプテスト
-        message_types = [
-            ("info", "Information message"),
-            ("warning", "Warning message"),
-            ("error", "Error message"),
-            ("success", "Success message"),
-            ("debug", "Debug message"),
-        ]
-
-        for msg_type, content in message_types:
-            method_names = [
-                f"show_{msg_type}",
-                f"display_{msg_type}",
-                f"print_{msg_type}",
-                msg_type,
-            ]
-
-            for method_name in method_names:
-                if hasattr(messaging, method_name):
-                    method = getattr(messaging, method_name)
-                    if callable(method):
-                        try:
-                            method(content)
-                        except Exception:
-                            pass
-
-    def test_console_messaging_formatting(self):
-        """メッセージフォーマット機能テスト"""
-        from kumihan_formatter.ui.console_messaging import ConsoleMessaging
-
-        messaging = ConsoleMessaging()
-
-        # フォーマット機能テスト
-        format_methods = [
-            "format_message",
-            "format_error",
-            "format_warning",
-            "format_success",
-            "add_color",
-            "add_style",
-        ]
-
-        for method_name in format_methods:
-            if hasattr(messaging, method_name):
-                method = getattr(messaging, method_name)
-                if callable(method):
-                    try:
-                        result = method("Test message")
-                        assert isinstance(result, str)
-                    except Exception:
-                        try:
-                            result = method("Test", "type")
-                            assert isinstance(result, str)
-                        except:
-                            pass
-
-    def test_console_messaging_templates(self):
-        """メッセージテンプレート機能テスト"""
-        from kumihan_formatter.ui.console_messaging import ConsoleMessaging
-
-        messaging = ConsoleMessaging()
-
-        # テンプレート機能テスト
-        template_methods = [
-            "apply_template",
-            "use_template",
-            "format_with_template",
-            "get_template",
-            "set_template",
-        ]
-
-        for method_name in template_methods:
-            if hasattr(messaging, method_name):
-                method = getattr(messaging, method_name)
-                if callable(method):
-                    try:
-                        if "get" in method_name:
-                            result = method("default")
-                        elif "set" in method_name:
-                            method("default", "template content")
-                        else:
-                            result = method("message", "template")
-                            if result is not None:
-                                assert isinstance(result, str)
-                    except Exception:
-                        pass
