--- tests/test_core_coverage_boost.py
@@ -1,414 +1,0 @@
-"""Core Coverage Boost Tests
-
-Phase 2 tests to push coverage from 11% to 30-50%.
-Focus on high-impact core modules.
-"""
-
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-
-class TestRenderingCoverageBoosting:
-    """Boost rendering module coverage significantly"""
-
-    def test_main_renderer_comprehensive(self):
-        """Test main renderer comprehensive functionality"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        renderer = HTMLRenderer()
-
-        # Test different node types rendering
-        test_nodes = [
-            Node("p", "Simple paragraph"),
-            Node("h1", "Main heading"),
-            Node("h2", "Sub heading"),
-            Node("strong", "Bold text"),
-            Node("em", "Italic text"),
-            Node("div", "Container"),
-        ]
-
-        for node in test_nodes:
-            try:
-                result = renderer.render_node(node)
-                assert isinstance(result, str)
-                assert len(result) > 0
-            except Exception:
-                # Some methods may not be fully implemented
-                pass
-
-        # Test list rendering
-        try:
-            result = renderer.render_nodes(test_nodes)
-            assert isinstance(result, str)
-            assert len(result) > 0
-        except Exception:
-            pass
-
-        # Test nesting order
-        assert hasattr(renderer, "NESTING_ORDER")
-        assert isinstance(renderer.NESTING_ORDER, list)
-        assert len(renderer.NESTING_ORDER) > 0
-
-    def test_element_renderer_functionality(self):
-        """Test element renderer core functions"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.core.rendering.element_renderer import ElementRenderer
-
-        renderer = ElementRenderer()
-
-        # Test basic element rendering
-        test_cases = [
-            ("p", "text", "<p>text</p>"),
-            ("h1", "heading", "<h1>heading</h1>"),
-            ("strong", "bold", "<strong>bold</strong>"),
-            ("em", "italic", "<em>italic</em>"),
-        ]
-
-        for node_type, content, expected_pattern in test_cases:
-            node = Node(node_type, content)
-            try:
-                result = renderer.render_element(node)
-                assert isinstance(result, str)
-                assert content in result
-            except Exception:
-                # Some methods may not be implemented
-                pass
-
-        # Test attribute rendering
-        node_with_attrs = Node("div", "content")
-        node_with_attrs.add_attribute("class", "test-class")
-        node_with_attrs.add_attribute("id", "test-id")
-
-        try:
-            result = renderer.render_element(node_with_attrs)
-            assert isinstance(result, str)
-            # Should contain attributes in some form
-        except Exception:
-            pass
-
-
-class TestParsingCoverageBoosting:
-    """Boost parsing module coverage significantly"""
-
-    def test_block_parser_comprehensive(self):
-        """Test block parser comprehensive functionality"""
-        from kumihan_formatter.core.block_parser.block_parser import BlockParser
-
-        try:
-            from kumihan_formatter.core.keyword_parsing.definitions import (
-                KeywordDefinitions,
-            )
-
-            # KeywordParserが存在しない場合はMarkerParserを使用
-            from kumihan_formatter.core.keyword_parsing.marker_parser import (
-                MarkerParser,
-            )
-
-            # 必要な依存関係を作成
-            keyword_definitions = KeywordDefinitions()
-            keyword_parser = MarkerParser(keyword_definitions)
-        except ImportError:
-            pytest.skip("Required keyword parsing modules not available")
-        parser = BlockParser(keyword_parser)
-
-        # Test different block types
-        test_blocks = [
-            "Simple paragraph text",
-            "# Heading 1",
-            "## Heading 2",
-            "**Bold text**",
-            "*Italic text*",
-            ";;;highlight;;; highlighted text ;;;",
-            ";;;box;;; boxed content ;;;",
-        ]
-
-        for block_text in test_blocks:
-            try:
-                result = parser.parse_block(block_text)
-                assert result is not None
-            except Exception:
-                # Some parsing may fail due to incomplete implementation
-                pass
-
-        # Test line parsing
-        try:
-            lines = ["Line 1", "Line 2", "Line 3"]
-            result = parser.parse_lines(lines)
-            assert result is not None
-        except Exception:
-            pass
-
-    def test_keyword_parser_comprehensive(self):
-        """Test keyword parser comprehensive functionality"""
-        from kumihan_formatter.core.keyword_parser import KeywordParser
-
-        parser = KeywordParser()
-
-        # Test keyword detection
-        test_texts = [
-            ";;;highlight;;; test content ;;;",
-            ";;;box;;; boxed content ;;;",
-            ";;;footnote;;; footnote text ;;;",
-            "((footnote content))",
-            "｜ruby《reading》",
-        ]
-
-        for text in test_texts:
-            try:
-                result = parser.parse(text)
-                assert result is not None
-            except Exception:
-                # Some parsing may not be fully implemented
-                pass
-
-        # Test keyword validation
-        try:
-            valid_keywords = parser.get_valid_keywords()
-            assert isinstance(valid_keywords, (list, set, tuple))
-        except Exception:
-            pass
-
-    def test_marker_parser_comprehensive(self):
-        """Test marker parser comprehensive functionality"""
-        from kumihan_formatter.core.keyword_parsing.definitions import (
-            KeywordDefinitions,
-        )
-        from kumihan_formatter.core.keyword_parsing.marker_parser import MarkerParser
-
-        # 必要な依存関係を作成
-        keyword_definitions = KeywordDefinitions()
-        parser = MarkerParser(keyword_definitions)
-
-        # Test marker detection
-        test_markers = [
-            ";;;highlight;;;",
-            ";;;box;;;",
-            ";;;footnote;;;",
-            "(((",
-            ")))",
-            "｜",
-            "《",
-            "》",
-        ]
-
-        for marker in test_markers:
-            try:
-                result = parser.detect_marker(marker)
-                assert result is not None
-            except Exception:
-                pass
-
-        # Test marker parsing
-        test_texts = [
-            ";;;highlight;;; content ;;;",
-            "((footnote))",
-            "｜ruby《reading》",
-        ]
-
-        for text in test_texts:
-            try:
-                result = parser.parse_markers(text)
-                assert result is not None
-            except Exception:
-                pass
-
-
-class TestUtilitiesCoverageBoosting:
-    """Boost utilities module coverage significantly"""
-
-    def test_logger_comprehensive(self):
-        """Test logger comprehensive functionality"""
-        from kumihan_formatter.core.utilities.logger import get_logger
-
-        # Test basic logger functionality
-        logger = get_logger("test_module")
-        assert logger is not None
-
-        # Test different log levels
-        log_methods = ["debug", "info", "warning", "error", "critical"]
-        for method_name in log_methods:
-            if hasattr(logger, method_name):
-                method = getattr(logger, method_name)
-                try:
-                    method("Test message")
-                except Exception:
-                    # Logging may fail in test environment
-                    pass
-
-        # Test logger configuration
-        try:
-            logger.setLevel("INFO")
-        except Exception:
-            pass
-
-    def test_structured_logger_base_comprehensive(self):
-        """Test structured logger base comprehensive functionality"""
-        import logging
-
-        from kumihan_formatter.core.utilities.structured_logger_base import (
-            StructuredLogger,
-        )
-
-        base_logger = logging.getLogger("test_structured")
-        structured_logger = StructuredLogger(base_logger)
-
-        # Test structured logging methods
-        test_contexts = [
-            {"operation": "test", "status": "success"},
-            {"file_path": "/test/path", "size": 1024},
-            {"error_type": "ValidationError", "line": 42},
-        ]
-
-        for context in test_contexts:
-            try:
-                structured_logger.log_with_context("INFO", "Test message", **context)
-            except Exception:
-                pass
-
-        # Test performance logging
-        try:
-            structured_logger.log_performance("test_operation", 0.05, iterations=100)
-        except Exception:
-            pass
-
-        # Test error logging with suggestions
-        try:
-            structured_logger.log_error_with_suggestion(
-                "Test error", "Try this fix", error_type="TestError"
-            )
-        except Exception:
-            pass
-
-    def test_text_processor_comprehensive(self):
-        """Test text processor comprehensive functionality"""
-        from kumihan_formatter.core.utilities.text_processor import TextProcessor
-
-        processor = TextProcessor()
-
-        # Test text processing methods
-        test_texts = [
-            "  hello   world  ",
-            "Text\n\nwith\n\nmultiple\n\nlines",
-            "<p>HTML content</p>",
-            "Special & chars < >",
-            "Japanese　text　with　full-width　spaces",
-        ]
-
-        for text in test_texts:
-            # Test normalization
-            try:
-                result = processor.normalize_whitespace(text)
-                assert isinstance(result, str)
-            except Exception:
-                pass
-
-            # Test cleaning
-            try:
-                result = processor.clean_text(text)
-                assert isinstance(result, str)
-            except Exception:
-                pass
-
-            # Test processing
-            try:
-                result = processor.process_text(text)
-                assert isinstance(result, str)
-            except Exception:
-                pass
-
-
-class TestNodeCoverageBoosting:
-    """Boost AST Node module coverage significantly"""
-
-    def test_node_comprehensive_functionality(self):
-        """Test Node class comprehensive functionality"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-
-        # Test different node types and content
-        test_cases = [
-            ("p", "Simple text"),
-            ("h1", "Heading"),
-            ("div", ["nested", "content"]),
-            ("ul", [Node("li", "Item 1"), Node("li", "Item 2")]),
-            ("details", Node("p", "Detail content")),
-        ]
-
-        for node_type, content in test_cases:
-            node = Node(node_type, content)
-
-            # Test type checking methods
-            assert isinstance(node.is_block_element(), bool)
-            assert isinstance(node.is_inline_element(), bool)
-            assert isinstance(node.is_list_element(), bool)
-            assert isinstance(node.is_heading(), bool)
-
-            # Test heading level functionality
-            heading_level = node.get_heading_level()
-            if node.is_heading():
-                assert heading_level is not None
-                assert 1 <= heading_level <= 5
-            else:
-                assert heading_level is None
-
-            # Test content checking
-            assert isinstance(node.contains_text(), bool)
-            text_content = node.get_text_content()
-            assert isinstance(text_content, str)
-
-            # Test child operations
-            child_count = node.count_children()
-            assert isinstance(child_count, int)
-            assert child_count >= 0
-
-            # Test attribute operations
-            node.add_attribute("test_attr", "test_value")
-            assert node.has_attribute("test_attr")
-            assert node.get_attribute("test_attr") == "test_value"
-            assert node.get_attribute("nonexistent", "default") == "default"
-
-        # Test node tree walking
-        root = Node(
-            "div",
-            [
-                Node("h1", "Title"),
-                Node("p", "Paragraph"),
-                Node("ul", [Node("li", "Item")]),
-            ],
-        )
-
-        walked_nodes = list(root.walk())
-        assert len(walked_nodes) > 1
-        assert all(isinstance(n, Node) for n in walked_nodes)
-
-        # Test finding children by type
-        list_children = root.find_children_by_type("ul")
-        assert isinstance(list_children, list)
-
-    def test_node_edge_cases(self):
-        """Test Node class edge cases and error handling"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-
-        # Test empty content
-        empty_node = Node("p", "")
-        assert not empty_node.contains_text()
-        assert empty_node.get_text_content() == ""
-        assert empty_node.count_children() == 0
-
-        # Test None content handling
-        try:
-            none_node = Node("p", None)
-            assert none_node.get_text_content() == ""
-        except Exception:
-            # May not handle None gracefully
-            pass
-
-        # Test invalid heading levels
-        invalid_headings = ["h0", "h6", "h7", "heading", "p"]
-        for heading_type in invalid_headings:
-            node = Node(heading_type, "Test")
-            level = node.get_heading_level()
-            if heading_type in ["h0", "h6", "h7"]:
-                assert level is None  # Invalid levels should return None
