+++ tests/test_core_integration_basic.py
@@ -211,188 +211,3 @@ class TestCoreIntegration:
             )
             assert callable(getattr(context, "custom", None))
             assert callable(getattr(context, "get", None))
-
-
-class TestParserRendererIntegration:
-    """Integration tests for parser-renderer workflow"""
-
-    def test_parse_render_workflow(self):
-        """Test complete parse-render workflow"""
-        # Simple workflow test
-        content = "Hello World"
-
-        try:
-            # Parse content
-            parser = Parser()
-            parsed_result = parser.parse(content)
-
-            # Render result
-            renderer = Renderer()
-            if parsed_result:
-                rendered_result = renderer.render(parsed_result)
-                assert isinstance(rendered_result, str)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            # If full workflow needs complex setup, just verify classes exist
-            pytest.skip(
-                f"Dependency unavailable: {type(e).__name__}: Module not available for testing"
-            )
-            assert Parser is not None
-            assert Renderer is not None
-
-    def test_function_based_workflow(self):
-        """Test function-based parse-render workflow"""
-        content = "Hello World"
-
-        try:
-            # Parse and render using functions
-            parsed = parse(content)
-            if parsed:
-                rendered = render(parsed)
-                assert isinstance(rendered, str)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            # Verify functions are callable
-            pytest.skip(
-                f"Dependency unavailable: {type(e).__name__}: Module not available for testing"
-            )
-            assert callable(parse)
-            assert callable(render)
-
-
-class TestFileOperationsIntegration:
-    """Integration tests for file operations"""
-
-    def test_file_read_write_cycle(self):
-        """Test file read-write cycle"""
-        file_ops = FileOperations()
-
-        # Create temporary file
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".txt", delete=False) as f:
-            test_content = "Test content for file operations"
-            f.write(test_content)
-            temp_path = f.name
-
-        try:
-            # Test file reading (actual API)
-            if hasattr(file_ops, "read_text_file"):
-                read_content = file_ops.read_text_file(temp_path)
-                if read_content is not None:
-                    assert isinstance(read_content, str)
-
-            # Test file writing (actual API)
-            if hasattr(file_ops, "write_text_file"):
-                new_content = "New test content"
-                result = file_ops.write_text_file(temp_path, new_content)
-                # Verify write operation completed (result might be bool or None)
-                assert result is not None or result is None
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            # Verify methods exist even if they need specific setup
-            pytest.skip(
-                f"Dependency unavailable: {type(e).__name__}: Module not available for testing"
-            )
-            assert hasattr(file_ops, "read_text_file")
-            assert hasattr(file_ops, "write_text_file")
-        finally:
-            # Cleanup
-            Path(temp_path).unlink(missing_ok=True)
-
-
-class TestConfigurationIntegration:
-    """Integration tests for configuration management"""
-
-    def test_config_template_integration(self):
-        """Test configuration with template system"""
-        config = ConfigManager()
-        template_manager = TemplateManager()
-
-        # Test that both systems can work together
-        assert config is not None
-        assert template_manager is not None
-
-        # Test config provides data for templates
-        config_dict = config.to_dict()
-        assert isinstance(config_dict, dict)
-
-        # Test render context can use config data
-        context = RenderContext()
-        try:
-            for key, value in config_dict.items():
-                context.custom(f"config_{key}", value)
-
-            # Verify variables were set
-            test_key = list(config_dict.keys())[0] if config_dict else "theme_name"
-            stored_value = context.get(f"config_{test_key}")
-            assert stored_value is not None or stored_value is None  # Either works
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            # Verify basic integration is possible
-            pytest.skip(
-                f"Dependency unavailable: {type(e).__name__}: Module not available for testing"
-            )
-            assert hasattr(config, "to_dict")
-            assert hasattr(context, "custom")
-
-    def test_config_parser_integration(self):
-        """Test configuration with parser system"""
-        config = ConfigManager()
-        parser = KeywordParser(config)
-
-        # Test parser can use config
-        assert parser is not None
-        assert hasattr(parser, "BLOCK_KEYWORDS")
-
-        # Test config influences parser behavior
-        keywords = parser.BLOCK_KEYWORDS
-        assert isinstance(keywords, dict)
-        assert len(keywords) > 0
-
-
-class TestASTNodeIntegration:
-    """Integration tests for AST node operations"""
-
-    def test_node_builder_workflow(self):
-        """Test complete node building workflow"""
-        # Test creating nodes with builder
-        builder = NodeBuilder("div")
-        node = builder.content("Test content").css_class("test-class").build()
-
-        assert node.type == "div"
-        assert node.content == "Test content"
-        assert "class" in node.attributes
-        assert node.attributes["class"] == "test-class"
-
-    def test_nested_node_creation(self):
-        """Test creating nested node structures"""
-        # Create parent node
-        parent = NodeBuilder("div").css_class("parent").build()
-
-        # Create child node
-        child = NodeBuilder("p").content("Child content").build()
-
-        # Test node relationships
-        parent.content = [child]
-        assert isinstance(parent.content, list)
-        assert len(parent.content) == 1
-        assert parent.content[0] == child
