--- tests/test_error_stream_config.py
@@ -1,315 +1,0 @@
-"""Error Handling, Stream Processing and Config File Tests
-
-Focus on file I/O, encoding handling, and related functionality.
-Target: Increase file processing module coverage significantly.
-"""
-
-import io
-import tempfile
-from pathlib import Path
-
-import pytest
-
-
-class TestErrorHandling:
-    """Test error handling in file operations"""
-
-    def test_file_not_found_handling(self):
-        """Test handling of missing files"""
-        try:
-            from kumihan_formatter.core.file_converter import FileConverter
-
-            converter = FileConverter()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Test non-existent input file
-            try:
-                converter.convert_file("non_existent_file.txt", "output.html")
-                # Should raise an exception or handle gracefully
-            except (FileNotFoundError, IOError):
-                # Expected behavior
-                pass
-            except AttributeError as ae:
-                # Method might not exist
-                pass
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_permission_error_handling(self):
-        """Test handling of permission errors"""
-        try:
-            from kumihan_formatter.core.file_converter import FileConverter
-
-            converter = FileConverter()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Test permission error (try to write to root directory)
-            try:
-                converter.convert_file(__file__, "/root/output.html")
-                # Should handle permission error gracefully
-            except (PermissionError, IOError):
-                # Expected behavior
-                pass
-            except AttributeError as ae:
-                # Method might not exist
-                pass
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_encoding_error_handling(self):
-        """Test handling of encoding errors"""
-        try:
-            from kumihan_formatter.core.utilities.file_reader import FileReader
-
-            reader = FileReader()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Create file with problematic encoding
-            with tempfile.NamedTemporaryFile(
-                mode="wb", suffix=".txt", delete=False
-            ) as f:
-                f.write(b"\xff\xfe\xff\xff")  # Invalid UTF-8
-                temp_path = f.name
-
-            try:
-                # Test reading problematic file
-                content = reader.read_file(temp_path)
-                # Should handle encoding errors gracefully
-                assert isinstance(content, str)
-
-            except (UnicodeDecodeError, UnicodeError):
-                # Expected behavior
-                pass
-            except AttributeError as ae:
-                # Method might not exist
-                pass
-            finally:
-                Path(temp_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-
-class TestFileSystemOperations:
-    """Test file system operations"""
-
-    def test_directory_operations(self):
-        """Test directory creation and management"""
-        try:
-            from kumihan_formatter.core.utilities.file_utils import FileUtils
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-            utils = FileUtils()
-
-            # Test directory creation
-            test_dir = Path(tempfile.gettempdir()) / "test_kumihan"
-
-            try:
-                utils.create_directory(str(test_dir))
-                assert test_dir.exists()
-                assert test_dir.is_dir()
-
-                # Test directory cleanup
-                utils.remove_directory(str(test_dir))
-                assert not test_dir.exists()
-
-            except AttributeError as ae:
-                # Methods might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_file_backup_operations(self):
-        """Test file backup functionality"""
-        try:
-            from kumihan_formatter.core.utilities.backup_manager import BackupManager
-
-            backup_manager = BackupManager()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Create test file
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".txt", delete=False
-            ) as f:
-                f.write("Original content")
-                original_path = f.name
-
-            try:
-                # Test backup creation
-                backup_path = backup_manager.create_backup(original_path)
-                assert Path(backup_path).exists()
-
-                # Test backup restoration
-                backup_manager.restore_backup(backup_path, original_path)
-
-                # Verify restoration
-                with open(original_path, "r", encoding="utf-8") as f:
-                    restored_content = f.read()
-                    assert "Original content" in restored_content
-
-            except AttributeError as ae:
-                # Methods might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-            finally:
-                Path(original_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-
-class TestStreamProcessing:
-    """Test stream-based processing"""
-
-    def test_stream_parser(self):
-        """Test stream-based parsing"""
-        try:
-            from kumihan_formatter.core.stream_parser import StreamParser
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-            parser = StreamParser()
-
-            # Test string stream
-            content = "# Heading\n\nParagraph content."
-            stream = io.StringIO(content)
-
-            try:
-                result = parser.parse_stream(stream)
-                assert result is not None
-
-            except AttributeError as ae:
-                # Method might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_stream_renderer(self):
-        """Test stream-based rendering"""
-        from kumihan_formatter.core.ast_nodes.node import Node
-
-        try:
-            from kumihan_formatter.core.stream_renderer import StreamRenderer
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-            renderer = StreamRenderer()
-
-            # Test rendering to stream
-            nodes = [Node("p", "Test content")]
-            output_stream = io.StringIO()
-
-            try:
-                renderer.render_to_stream(nodes, output_stream)
-                result = output_stream.getvalue()
-                assert isinstance(result, str)
-                assert "Test content" in result
-
-            except AttributeError as ae:
-                # Method might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-
-class TestConfigFileProcessing:
-    """Test configuration file processing"""
-
-    def test_config_file_loading(self):
-        """Test loading configuration from files"""
-        try:
-            from kumihan_formatter.config.config_loader import ConfigLoader
-
-            loader = ConfigLoader()
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-
-            # Test JSON config
-            config_data = {
-                "output_format": "html",
-                "encoding": "utf-8",
-                "template": "default",
-            }
-
-            with tempfile.NamedTemporaryFile(
-                mode="w", suffix=".json", delete=False
-            ) as f:
-                import json
-
-                json.dump(config_data, f)
-                config_path = f.name
-
-            try:
-                loaded_config = loader.load_json(config_path)
-                assert isinstance(loaded_config, dict)
-                assert loaded_config["output_format"] == "html"
-
-            except AttributeError as ae:
-                # Method might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-            finally:
-                Path(config_path).unlink(missing_ok=True)
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-
-    def test_config_file_validation(self):
-        """Test configuration file validation"""
-        try:
-            from kumihan_formatter.config.config_validator import ConfigValidator
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
-            return
-
-        try:
-            validator = ConfigValidator()
-
-            # Test valid config
-            valid_config = {"output_format": "html", "encoding": "utf-8"}
-
-            try:
-                is_valid = validator.validate_config(valid_config)
-                assert isinstance(is_valid, bool)
-
-                # Test invalid config
-                invalid_config = {"output_format": "unknown_format"}
-
-                is_invalid = validator.validate_config(invalid_config)
-                # Result depends on implementation
-
-            except AttributeError as ae:
-                # Methods might not exist
-                pytest.skip(f"Dependency unavailable: {type(ae).__name__}: {ae}")
-
-        except ImportError as ie:
-            pytest.skip(f"Dependency unavailable: {type(ie).__name__}: {ie}")
