--- tests/test_gui_models_threadsafe.py
@@ -1,342 +1,0 @@
-"""Thread-Safe GUI Models Tests
-
-Issue #516 Phase 5A対応 - GUI ModelsのThread-Safe設計とエラーハンドリングのテスト
-TDD実践による品質保証
-"""
-
-import tempfile
-import threading
-import time
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-from kumihan_formatter.gui_models import (
-    AppState,
-    ConversionState,
-    FileManager,
-    LogManager,
-)
-
-pytestmark = [pytest.mark.unit, pytest.mark.tdd_green]
-
-
-@pytest.mark.tdd_green
-class TestConversionStateThreadSafe:
-    """ConversionStateのThread-Safe機能テスト"""
-
-    def test_thread_safe_progress_setting(self):
-        """進捗設定のThread-Safe動作テスト"""
-        state = ConversionState()
-        results = []
-        errors = []
-
-        def update_progress(thread_id):
-            try:
-                for i in range(10):
-                    progress = thread_id * 10 + i
-                    if progress <= 100:
-                        state.set_progress(progress)
-                        results.append((thread_id, progress))
-                        time.sleep(0.001)  # 短い待機でタイミング調整
-            except Exception as e:
-                errors.append((thread_id, e))
-
-        # 複数スレッドで同時実行
-        threads = []
-        for i in range(5):
-            thread = threading.Thread(target=update_progress, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        # 全スレッド完了待機
-        for thread in threads:
-            thread.join()
-
-        # 結果確認
-        assert len(errors) == 0, f"エラーが発生: {errors}"
-        assert len(results) > 0, "結果が記録されていない"
-
-        # 最終進捗値が有効範囲内にあることを確認
-        final_progress = state.get_progress()
-        assert 0 <= final_progress <= 100
-
-    def test_callback_functionality(self):
-        """コールバック機能のテスト"""
-        state = ConversionState()
-        callback_calls = []
-
-        def test_callback(progress, status):
-            callback_calls.append((progress, status))
-
-        state.set_callback(test_callback)
-        state.set_progress(50)
-        state.set_status("テスト中")
-
-        assert len(callback_calls) >= 1
-        assert any(call[0] == 50 for call in callback_calls)
-
-    def test_elapsed_time_tracking(self):
-        """経過時間追跡のテスト"""
-        state = ConversionState()
-
-        # 処理開始
-        state.start_processing()
-        start_time = time.time()
-
-        # 短い待機
-        time.sleep(0.1)
-
-        # 経過時間確認
-        elapsed = state.get_elapsed_time()
-        actual_elapsed = time.time() - start_time
-
-        # 許容誤差内で経過時間が正確であることを確認
-        assert abs(elapsed - actual_elapsed) < 0.05
-
-    @pytest.mark.tdd_refactor
-    def test_error_handling_in_progress_setting(self):
-        """進捗設定時のエラーハンドリングテスト"""
-        state = ConversionState()
-
-        # 無効な値での設定（エラーが発生するが継続する）
-        with patch("logging.warning") as mock_warning:
-            state.set_progress("invalid")  # 文字列を設定
-            mock_warning.assert_called()
-
-        # 範囲外の値
-        with patch("logging.warning") as mock_warning:
-            state.set_progress(150)  # 100を超える値
-            mock_warning.assert_called()
-
-        # 負の値
-        with patch("logging.warning") as mock_warning:
-            state.set_progress(-10)
-            mock_warning.assert_called()
-
-
-@pytest.mark.tdd_green
-class TestFileManagerThreadSafe:
-    """FileManagerのThread-Safe機能テスト"""
-
-    def test_thread_safe_file_operations(self):
-        """ファイル操作のThread-Safe動作テスト"""
-        manager = FileManager()
-        results = []
-        errors = []
-
-        def test_file_operations(thread_id):
-            try:
-                with tempfile.TemporaryDirectory() as temp_dir:
-                    test_file = Path(temp_dir) / f"test_{thread_id}.txt"
-                    test_file.write_text(f"test content {thread_id}", encoding="utf-8")
-
-                    # ファイル情報取得
-                    info = manager.get_file_info(str(test_file))
-                    if info:
-                        results.append((thread_id, info))
-            except Exception as e:
-                errors.append((thread_id, e))
-
-        # 複数スレッドで同時実行
-        threads = []
-        for i in range(3):
-            thread = threading.Thread(target=test_file_operations, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        # 全スレッド完了待機
-        for thread in threads:
-            thread.join()
-
-        # 結果確認
-        assert len(errors) == 0, f"エラーが発生: {errors}"
-        assert len(results) == 3, "全スレッドの結果が記録されていない"
-
-    def test_file_cache_thread_safety(self):
-        """ファイルキャッシュのThread-Safe動作テスト"""
-        manager = FileManager()
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            test_file = Path(temp_dir) / "cache_test.txt"
-            test_file.write_text("test content", encoding="utf-8")
-
-            # 複数スレッドで同じファイルにアクセス
-            results = []
-
-            def access_file_info():
-                info = manager.get_file_info(str(test_file))
-                results.append(info)
-
-            threads = []
-            for _ in range(5):
-                thread = threading.Thread(target=access_file_info)
-                threads.append(thread)
-                thread.start()
-
-            for thread in threads:
-                thread.join()
-
-            # 全て同じ結果が返されることを確認
-            assert len(results) == 5
-            assert all(result is not None for result in results)
-            # キャッシュにより同じ情報が返されることを確認
-            first_result = results[0]
-            assert all(result["path"] == first_result["path"] for result in results)
-
-    def test_directory_validation_error_handling(self):
-        """ディレクトリ検証のエラーハンドリングテスト"""
-        # 無効なパスでのテスト
-        with patch("logging.warning") as mock_warning:
-            result = FileManager.validate_directory_writable("")
-            assert result is False
-
-        with patch("logging.warning") as mock_warning:
-            result = FileManager.validate_directory_writable(None)
-            assert result is False
-
-
-@pytest.mark.tdd_green
-class TestLogManagerThreadSafe:
-    """LogManagerのThread-Safe機能テスト"""
-
-    def test_concurrent_message_adding(self):
-        """並行メッセージ追加のテスト"""
-        log_manager = LogManager(max_messages=100)
-        results = []
-
-        def add_messages(thread_id):
-            for i in range(10):
-                message = f"Thread {thread_id} Message {i}"
-                formatted = log_manager.add_message(message, "info")
-                results.append((thread_id, formatted))
-
-        # 複数スレッドで同時実行
-        threads = []
-        for i in range(5):
-            thread = threading.Thread(target=add_messages, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join()
-
-        # 結果確認
-        assert len(results) == 50  # 5スレッド × 10メッセージ
-        assert log_manager.get_message_count() == 50
-
-    def test_message_limit_enforcement(self):
-        """メッセージ数制限の動作テスト"""
-        log_manager = LogManager(max_messages=10)
-
-        # 制限を超えるメッセージを追加
-        for i in range(20):
-            log_manager.add_message(f"Message {i}", "info")
-
-        # メッセージ数が制限内に収まっていることを確認
-        assert log_manager.get_message_count() <= 10
-
-    def test_level_based_message_retrieval(self):
-        """レベル別メッセージ取得のテスト"""
-        log_manager = LogManager()
-
-        # 異なるレベルのメッセージを追加
-        log_manager.add_message("Info message", "info")
-        log_manager.add_message("Error message", "error")
-        log_manager.add_message("Warning message", "warning")
-
-        # レベル別取得
-        error_messages = log_manager.get_messages_by_level("error")
-        info_messages = log_manager.get_messages_by_level("info")
-
-        assert len(error_messages) == 1
-        assert len(info_messages) == 1
-        assert error_messages[0]["message"] == "Error message"
-
-
-@pytest.mark.tdd_green
-class TestAppStateThreadSafe:
-    """AppStateのThread-Safe機能テスト"""
-
-    def test_conversion_readiness_check_thread_safety(self):
-        """変換準備チェックのThread-Safe動作テスト"""
-        app_state = AppState()
-        results = []
-
-        def check_readiness(thread_id):
-            ready, message = app_state.is_ready_for_conversion()
-            results.append((thread_id, ready, message))
-
-        # 複数スレッドで同時実行
-        threads = []
-        for i in range(3):
-            thread = threading.Thread(target=check_readiness, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join()
-
-        # 結果確認
-        assert len(results) == 3
-        # 全て同じ結果が返されることを確認（状態が変わらない限り）
-        assert all(isinstance(result[1], bool) for result in results)
-
-    def test_error_tracking_thread_safety(self):
-        """エラー追跡のThread-Safe動作テスト"""
-        app_state = AppState()
-        errors = []
-
-        def trigger_error(thread_id):
-            try:
-                # 無効な設定でエラーを発生させる
-                app_state.config.set_input_file("")  # 空のファイルパス
-                app_state.is_ready_for_conversion()  # エラーが記録される
-            except Exception as e:
-                errors.append((thread_id, e))
-
-        # 複数スレッドでエラーを発生
-        threads = []
-        for i in range(3):
-            thread = threading.Thread(target=trigger_error, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join(timeout=5)  # タイムアウト設定
-
-        # エラーが発生していないことを確認
-        assert len(errors) == 0, f"スレッドエラーが発生: {errors}"
-
-        # エラー情報確認
-        error_info = app_state.get_error_info()
-        assert error_info["has_errors"] is True
-        assert error_info["error_count"] > 0
-
-    @pytest.mark.tdd_refactor
-    def test_fallback_initialization(self):
-        """フォールバック初期化のテスト"""
-        try:
-            # 正常な初期化の確認
-            app_state = AppState()
-            assert hasattr(app_state, "config")
-            assert hasattr(app_state, "conversion_state")
-            assert hasattr(app_state, "file_manager")
-            assert hasattr(app_state, "log_manager")
-        except Exception as e:
-            pytest.skip(f"AppState初期化失敗: {e}")
-
-    def test_debug_mode_detection(self):
-        """デバッグモード検出のテスト"""
-        try:
-            # 環境変数なしの場合
-            app_state = AppState()
-            assert isinstance(app_state.debug_mode, bool)
-
-            # 環境変数ありの場合
-            with patch.dict("os.environ", {"KUMIHAN_GUI_DEBUG": "true"}):
-                app_state_debug = AppState()
-                assert app_state_debug.debug_mode is True
-        except Exception as e:
-            pytest.skip(f"デバッグモード検出テスト失敗: {e}")
