--- tests/test_import_data_filesystem.py
@@ -1,388 +1,0 @@
-"""Import, Data Structures and FileSystem Coverage Tests
-
-Targeted tests to push coverage from 12% toward 20-30%.
-Focus on previously untested but easily testable modules.
-"""
-
-import json
-import tempfile
-from pathlib import Path
-from unittest.mock import Mock, patch
-
-import pytest
-
-
-class TestImportCoverageBoost:
-    """Test imports to boost basic module coverage"""
-
-    def test_all_core_imports(self):
-        """Test importing all core modules for basic coverage"""
-        # Core modules - just importing them provides basic coverage
-        core_modules = [
-            "kumihan_formatter.core.ast_nodes",
-            "kumihan_formatter.core.rendering",
-            "kumihan_formatter.core.utilities",
-            "kumihan_formatter.core.file_operations",
-            "kumihan_formatter.config",
-            "kumihan_formatter.parser",
-            "kumihan_formatter.renderer",
-        ]
-
-        for module_name in core_modules:
-            try:
-                __import__(module_name)
-            except ImportError as e:
-                pytest.skip(f"Module not available: {e}")
-
-    def test_specific_module_instantiation(self):
-        """Test instantiating classes for basic coverage"""
-        instantiation_tests = [
-            ("kumihan_formatter.core.ast_nodes.node", "Node", ("p", "content")),
-            ("kumihan_formatter.core.ast_nodes.node_builder", "NodeBuilder", ("div",)),
-            ("kumihan_formatter.core.rendering.html_utils", "render_attributes", None),
-            ("kumihan_formatter.core.utilities.converters", "Converters", None),
-            ("kumihan_formatter.core.file_operations", "FileOperations", None),
-        ]
-
-        for module_name, class_name, args in instantiation_tests:
-            try:
-                module = __import__(module_name, fromlist=[class_name])
-                if hasattr(module, class_name):
-                    cls = getattr(module, class_name)
-                    if args is None:
-                        # Function or no-arg class
-                        if callable(cls):
-                            try:
-                                result = cls()
-                            except (
-                                AttributeError,
-                                NotImplementedError,
-                                TypeError,
-                                ValueError,
-                                FileNotFoundError,
-                            ) as e:
-                                pytest.skip(f"Method or operation not available: {e}")
-                    else:
-                        # Class with args
-                        try:
-                            instance = cls(*args)
-                            assert instance is not None
-                        except (
-                            AttributeError,
-                            NotImplementedError,
-                            TypeError,
-                            ValueError,
-                            FileNotFoundError,
-                        ) as e:
-                            pytest.skip(f"Method or operation not available: {e}")
-            except ImportError as e:
-                pytest.skip(f"Module not available: {e}")
-
-
-class TestDataStructuresCoverage:
-    """Test data structures for coverage"""
-
-    def test_data_structures_comprehensive(self):
-        """Test data structures comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.utilities.data_structures import (
-                Queue,
-                Stack,
-                TreeNode,
-            )
-
-            # DataStructuresクラスが存在しない場合は個別クラスを使用
-        except ImportError as e:
-            pytest.skip(f"DataStructures not available: {e}")
-            return
-
-        try:
-            # Create mock data structure handler
-            class MockDataStructures:
-                def validate(self, data):
-                    return True
-
-                def transform(self, data):
-                    return data
-
-                def serialize(self, data):
-                    return str(data)
-
-                def get_type(self, data):
-                    return type(data).__name__
-
-                def flatten(self, data):
-                    return data
-
-                def get_depth(self, data):
-                    return 1
-
-            ds = MockDataStructures()
-
-            # Test various data structure operations
-            test_data = [
-                {"key1": "value1", "key2": "value2"},
-                [1, 2, 3, 4, 5],
-                ("tuple", "data"),
-                "string data",
-                42,
-                3.14,
-            ]
-
-            for data in test_data:
-                try:
-                    # Test data validation
-                    is_valid = ds.validate(data)
-                    assert isinstance(is_valid, bool)
-
-                    # Test data transformation
-                    transformed = ds.transform(data)
-                    assert transformed is not None
-
-                    # Test data serialization
-                    serialized = ds.serialize(data)
-                    assert serialized is not None
-
-                    # Test data type detection
-                    data_type = ds.get_type(data)
-                    assert isinstance(data_type, str)
-
-                except (
-                    AttributeError,
-                    NotImplementedError,
-                    TypeError,
-                    ValueError,
-                    FileNotFoundError,
-                ) as e:
-                    pytest.skip(f"Method or operation not available: {e}")
-
-            # Test nested structures
-            nested_data = {
-                "level1": {"level2": {"level3": ["item1", "item2", "item3"]}},
-                "array": [{"id": 1}, {"id": 2}],
-            }
-
-            try:
-                flattened = ds.flatten(nested_data)
-                assert flattened is not None
-
-                depth = ds.get_depth(nested_data)
-                assert isinstance(depth, int)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-    def test_converters_comprehensive(self):
-        """Test converters comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.utilities.converters import (
-                convert_to_dict,
-                convert_to_tree,
-            )
-
-            # Convertersクラスが存在しない場合は個別関数を使用
-        except ImportError as e:
-            pytest.skip(f"Converters not available: {e}")
-            return
-
-        # Create mock Converters class
-        class Converters:
-            def convert(self, data_type, value, operation):
-                if operation == "upper":
-                    return str(value).upper()
-                elif operation == "lower":
-                    return str(value).lower()
-                else:
-                    return value
-
-        converters = Converters()
-
-        # Test various conversion scenarios
-        conversion_tests = [
-            # String conversions
-            ("string", "hello world", "upper"),
-            ("string", "HELLO WORLD", "lower"),
-            ("string", "hello_world", "camel_case"),
-            # Number conversions
-            ("number", "42", "int"),
-            ("number", "3.14", "float"),
-            ("number", 1024, "bytes"),
-            # Date/time conversions
-            ("datetime", "2024-01-15", "timestamp"),
-            ("datetime", "2024-01-15 10:30:00", "iso"),
-            # Data format conversions
-            ("format", {"key": "value"}, "json"),
-            ("format", ["item1", "item2"], "csv"),
-        ]
-
-        for conv_type, input_data, target_format in conversion_tests:
-            try:
-                result = converters.convert(input_data, target_format)
-                assert result is not None
-
-                # Test reverse conversion if available
-                if hasattr(converters, "reverse_convert"):
-                    reversed_result = converters.reverse_convert(result, conv_type)
-                    assert reversed_result is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-        # Test batch conversions
-        try:
-            batch_data = ["item1", "item2", "item3"]
-            batch_result = converters.batch_convert(batch_data, "upper")
-            assert isinstance(batch_result, list)
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            FileNotFoundError,
-        ) as e:
-            pytest.skip(f"Method or operation not available: {e}")
-
-
-class TestFileSystemCoverage:
-    """Test file system utilities for coverage"""
-
-    def test_file_system_comprehensive(self):
-        """Test file system comprehensive functionality"""
-        try:
-            from kumihan_formatter.core.utilities.file_system import (
-                ensure_directory,
-                get_file_info,
-            )
-
-            # FileSystemクラスが存在しない場合は個別関数を使用
-        except ImportError as e:
-            pytest.skip(f"FileSystem not available: {e}")
-            return
-
-        # Create mock FileSystem class
-        class FileSystem:
-            def list_files(self, path):
-                return [f"file_{i}.txt" for i in range(3)]
-
-            def get_info(self, path):
-                return {"size": 100, "type": "file"}
-
-            def create_directory(self, path):
-                return True
-
-        fs = FileSystem()
-
-        # Create test directory structure
-        with tempfile.TemporaryDirectory() as temp_dir:
-            temp_path = Path(temp_dir)
-
-            # Create test files and directories
-            test_files = []
-            for i in range(3):
-                file_path = temp_path / f"test_file_{i}.txt"
-                file_path.write_text(f"Content of file {i}")
-                test_files.append(str(file_path))
-
-            subdir = temp_path / "subdir"
-            subdir.mkdir()
-            subfile = subdir / "nested_file.txt"
-            subfile.write_text("Nested content")
-
-            # Test directory operations
-            try:
-                # List directory contents
-                contents = fs.list_directory(str(temp_path))
-                assert isinstance(contents, list)
-                assert len(contents) > 0
-
-                # Test recursive listing
-                recursive_contents = fs.list_recursive(str(temp_path))
-                assert isinstance(recursive_contents, list)
-                assert len(recursive_contents) >= len(contents)
-
-                # Test directory tree
-                tree = fs.get_directory_tree(str(temp_path))
-                assert tree is not None
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
-
-            # Test file operations
-            for file_path in test_files:
-                try:
-                    # Test file existence
-                    exists = fs.file_exists(file_path)
-                    assert exists is True
-
-                    # Test file stats
-                    stats = fs.get_file_stats(file_path)
-                    assert isinstance(stats, dict)
-
-                    # Test file permissions
-                    permissions = fs.get_permissions(file_path)
-                    assert permissions is not None
-
-                    # Test file type detection
-                    file_type = fs.get_file_type(file_path)
-                    assert isinstance(file_type, str)
-
-                except (
-                    AttributeError,
-                    NotImplementedError,
-                    TypeError,
-                    ValueError,
-                    FileNotFoundError,
-                ) as e:
-                    pytest.skip(f"Method or operation not available: {e}")
-
-            # Test path operations
-            try:
-                # Test path resolution
-                resolved = fs.resolve_path("./relative/path")
-                assert isinstance(resolved, str)
-
-                # Test path validation
-                is_valid = fs.validate_path(str(temp_path))
-                assert is_valid is True
-
-                # Test common path operations
-                parent = fs.get_parent(str(test_files[0]))
-                assert isinstance(parent, str)
-
-                basename = fs.get_basename(str(test_files[0]))
-                assert isinstance(basename, str)
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                FileNotFoundError,
-            ) as e:
-                pytest.skip(f"Method or operation not available: {e}")
