+++ tests/test_node_builder_deep_coverage.py
@@ -1,6 +1,6 @@
-"""Working Node Builder and Utilities Deep Coverage Tests
+"""Working Node Builder Deep Coverage Tests
 
-Focus on modules that are confirmed to be working and can be deeply tested.
+Focus on node builder functionality and comprehensive node method testing.
 Strategy: Exercise existing working code paths extensively.
 """
 
@@ -198,181 +198,3 @@ class TestWorkingNodeBuilderDeepCoverage:
 
             except Exception as e:
                 print(f"Node method test failed for {node_type}: {e}")
-
-
-class TestWorkingUtilitiesDeepCoverage:
-    """Deep test working utilities"""
-
-    def test_logger_comprehensive_all_scenarios(self):
-        """Test logger comprehensive all scenarios"""
-        from kumihan_formatter.core.utilities.logger import get_logger
-
-        # Test multiple logger instances with different configurations
-        logger_configs = [
-            ("main", {}),
-            ("parser", {"level": "DEBUG"}),
-            ("renderer", {"level": "INFO"}),
-            ("config", {"level": "WARNING"}),
-            ("file_ops", {"level": "ERROR"}),
-        ]
-
-        for logger_name, config in logger_configs:
-            try:
-                logger = get_logger(logger_name)
-                assert logger is not None
-
-                # Test all logging levels with various message types
-                test_messages = [
-                    "Simple message",
-                    "Message with data: %s",
-                    "Unicode message: 日本語",
-                    "Long message: " + "x" * 200,
-                    "",  # Empty message
-                ]
-
-                log_levels = ["debug", "info", "warning", "error", "critical"]
-
-                for level in log_levels:
-                    if hasattr(logger, level):
-                        log_method = getattr(logger, level)
-
-                        for message in test_messages:
-                            try:
-                                # Basic logging
-                                log_method(message)
-
-                                # Logging with formatting
-                                if "%s" in message:
-                                    log_method(message, "formatted_data")
-
-                                # Logging with extra data
-                                log_method(message, extra={"test_key": "test_value"})
-
-                            except Exception:
-                                pass
-
-                # Test logger configuration if available
-                if hasattr(logger, "setLevel"):
-                    logger.setLevel("INFO")
-
-                if hasattr(logger, "addFilter"):
-                    # Add a simple filter
-                    def test_filter(record):
-                        return True
-
-                    logger.addFilter(test_filter)
-
-            except Exception as e:
-                print(f"Logger test failed for {logger_name}: {e}")
-
-    def test_structured_logger_comprehensive_scenarios(self):
-        """Test structured logger comprehensive scenarios"""
-        import logging
-
-        from kumihan_formatter.core.utilities.structured_logger_base import (
-            StructuredLoggerBase,
-        )
-
-        # Create comprehensive structured logger test
-        base_logger = logging.getLogger("structured_comprehensive")
-        structured_logger = StructuredLoggerBase(base_logger)
-
-        # Test comprehensive structured logging scenarios
-        logging_scenarios = [
-            # Basic structured logging
-            {
-                "level": "INFO",
-                "message": "Basic operation",
-                "context": {"operation": "test", "status": "success"},
-            },
-            # Performance logging
-            {
-                "level": "INFO",
-                "message": "Performance data",
-                "context": {
-                    "operation": "parse",
-                    "duration_ms": 125.5,
-                    "items_processed": 150,
-                    "memory_usage": 1024 * 1024,
-                },
-            },
-            # Error logging with context
-            {
-                "level": "ERROR",
-                "message": "Operation failed",
-                "context": {
-                    "operation": "render",
-                    "error_type": "ValidationError",
-                    "file_path": "/test/file.txt",
-                    "line_number": 42,
-                },
-            },
-            # Security logging (with sensitive data filtering)
-            {
-                "level": "WARNING",
-                "message": "Security event",
-                "context": {
-                    "user": "test_user",
-                    "password": "secret123",  # Should be filtered
-                    "api_key": "abc123xyz",  # Should be filtered
-                    "action": "login_attempt",
-                    "ip_address": "192.168.1.1",
-                },
-            },
-            # Complex nested context
-            {
-                "level": "DEBUG",
-                "message": "Complex operation",
-                "context": {
-                    "request": {
-                        "method": "POST",
-                        "url": "/api/convert",
-                        "headers": {"Content-Type": "application/json"},
-                    },
-                    "response": {"status": 200, "size": 2048},
-                    "timing": {"start": "2024-01-15T10:30:00Z", "duration": 0.156},
-                },
-            },
-        ]
-
-        for scenario in logging_scenarios:
-            try:
-                level = scenario["level"]
-                message = scenario["message"]
-                context = scenario["context"]
-
-                # Test context logging
-                structured_logger.log_with_context(level, message, **context)
-
-                # Test specific logging methods if available
-                if (
-                    hasattr(structured_logger, "log_performance")
-                    and "duration_ms" in context
-                ):
-                    structured_logger.log_performance(
-                        context.get("operation", "test"), context["duration_ms"] / 1000
-                    )
-
-                if (
-                    hasattr(structured_logger, "log_error_with_context")
-                    and level == "ERROR"
-                ):
-                    structured_logger.log_error_with_context(message, **context)
-
-            except Exception as e:
-                print(f"Structured logging test failed for {scenario['level']}: {e}")
-
-        # Test log filtering and sanitization
-        try:
-            # This should filter out sensitive data
-            structured_logger.log_with_context(
-                "INFO",
-                "Test with sensitive data",
-                username="test",
-                password="should_be_filtered",
-                token="should_be_filtered",
-                safe_data="should_remain",
-            )
-
-        except Exception:
-            pass
