--- tests/test_realistic_config_coverage.py
@@ -1,369 +1,0 @@
-"""Realistic Config Coverage Tests
-
-Focus on actual working config functionality with realistic test cases.
-Target: Increase config module coverage significantly.
-"""
-
-import tempfile
-from pathlib import Path
-
-import pytest
-
-
-class TestSimpleConfig:
-    """Test SimpleConfig class"""
-
-    def test_simple_config_basic(self):
-        """Test basic SimpleConfig functionality"""
-        from kumihan_formatter.config import ConfigManager
-
-        config = ConfigManager()
-        assert config is not None
-
-        # Test basic config operations
-        try:
-            # Test setting values
-            config.set("test_key", "test_value")
-
-            # Test getting values
-            value = config.get("test_key")
-            assert value == "test_value"
-
-            # Test default values
-            default_value = config.get("nonexistent", "default")
-            assert default_value == "default"
-
-        except AttributeError:
-            # Methods might not exist, test alternative approach
-            try:
-                config.test_key = "test_value"
-                assert hasattr(config, "test_key")
-            except:
-                pass
-
-    def test_simple_config_css_operations(self):
-        """Test CSS-related operations"""
-        from kumihan_formatter.config import ConfigManager
-
-        config = ConfigManager()
-
-        # Test CSS configuration
-        css_settings = {
-            "font_family": "Arial, sans-serif",
-            "font_size": "14px",
-            "background_color": "#ffffff",
-            "text_color": "#000000",
-        }
-
-        for key, value in css_settings.items():
-            try:
-                config.set(key, value)
-                retrieved = config.get(key)
-                assert retrieved == value
-            except AttributeError:
-                # Alternative approach
-                try:
-                    setattr(config, key, value)
-                    retrieved = getattr(config, key, None)
-                    assert retrieved == value
-                except:
-                    pass
-
-    def test_simple_config_file_operations(self):
-        """Test file-related configuration"""
-        from kumihan_formatter.config import ConfigManager
-
-        config = ConfigManager()
-
-        # Test file settings
-        file_settings = {
-            "input_encoding": "utf-8",
-            "output_encoding": "utf-8",
-            "output_format": "html",
-            "template": "default",
-        }
-
-        for key, value in file_settings.items():
-            try:
-                config.set(key, value)
-                assert config.get(key) == value
-            except AttributeError:
-                pass
-
-
-class TestConfigManager:
-    """Test ConfigManager class"""
-
-    def test_config_manager_basic(self):
-        """Test basic ConfigManager functionality"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-        assert config_manager is not None
-
-        # Test basic operations
-        test_config = {
-            "parser": {"strict_mode": False, "encoding": "utf-8"},
-            "renderer": {"template": "default", "minify": False},
-        }
-
-        try:
-            config_manager.load_config(test_config)
-
-            # Test retrieval
-            parser_config = config_manager.get("parser")
-            assert parser_config is not None
-
-            renderer_config = config_manager.get("renderer")
-            assert renderer_config is not None
-
-        except AttributeError:
-            # Alternative testing
-            try:
-                config_manager.update(test_config)
-            except:
-                pass
-
-    def test_config_manager_validation(self):
-        """Test config validation"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-
-        # Test valid configuration
-        valid_config = {
-            "output_format": "html",
-            "encoding": "utf-8",
-            "template": "default",
-        }
-
-        try:
-            config_manager.load_config(valid_config)
-            is_valid = config_manager.validate()
-            assert isinstance(is_valid, bool)
-        except AttributeError:
-            # Validation might not be implemented
-            pass
-
-        # Test invalid configuration
-        invalid_config = {
-            "output_format": "unknown_format",
-            "encoding": "invalid_encoding",
-        }
-
-        try:
-            config_manager.load_config(invalid_config)
-            is_valid = config_manager.validate()
-            # Invalid config might still return True if validation is basic
-        except AttributeError:
-            pass
-
-
-class TestBaseConfig:
-    """Test BaseConfig class"""
-
-    def test_base_config_functionality(self):
-        """Test BaseConfig basic functionality"""
-        from kumihan_formatter.config.base_config import BaseConfig
-
-        config = BaseConfig()
-        assert config is not None
-
-        # Test configuration operations
-        test_data = {
-            "key1": "value1",
-            "key2": 42,
-            "key3": True,
-            "nested": {"subkey": "subvalue"},
-        }
-
-        for key, value in test_data.items():
-            try:
-                config.set(key, value)
-                retrieved = config.get(key)
-                assert retrieved == value
-            except AttributeError:
-                # Methods might not exist
-                pass
-
-    def test_base_config_nested_access(self):
-        """Test nested configuration access"""
-        from kumihan_formatter.config.base_config import BaseConfig
-
-        config = BaseConfig()
-
-        # Test nested configuration
-        nested_config = {"level1": {"level2": {"level3": "deep_value"}}}
-
-        try:
-            config.set("nested", nested_config)
-
-            # Test nested access
-            if hasattr(config, "get_nested"):
-                deep_value = config.get_nested("nested.level1.level2.level3")
-                assert deep_value == "deep_value"
-        except AttributeError:
-            pass
-
-
-class TestExtendedConfig:
-    """Test ExtendedConfig class"""
-
-    def test_extended_config_functionality(self):
-        """Test ExtendedConfig basic functionality"""
-        from kumihan_formatter.config.extended_config import ExtendedConfig
-
-        config = ExtendedConfig()
-        assert config is not None
-
-        # Test advanced configuration features
-        try:
-            # Test configuration validation
-            config.set("output_format", "html")
-            config.set("template", "default")
-
-            # Test validation
-            validation_result = config.validate_all()
-            assert isinstance(validation_result, (bool, dict, list))
-
-        except AttributeError:
-            pass
-
-    def test_extended_config_sections(self):
-        """Test configuration sections"""
-        from kumihan_formatter.config.extended_config import ExtendedConfig
-
-        config = ExtendedConfig()
-
-        # Test sections
-        sections = ["parser", "renderer", "output"]
-
-        for section in sections:
-            try:
-                config.create_section(section)
-                config.set_section_value(section, "enabled", True)
-
-                section_config = config.get_section(section)
-                assert isinstance(section_config, dict)
-
-            except AttributeError:
-                pass
-
-
-class TestConfigIntegration:
-    """Test config integration with other components"""
-
-    def test_config_with_parser(self):
-        """Test config integration with parser"""
-        from kumihan_formatter.config import ConfigManager
-        from kumihan_formatter.parser import Parser
-
-        parser = Parser()
-        config = ConfigManager()
-
-        # Test parser accepts config
-        try:
-            config.set("strict_mode", False)
-            config.set("encoding", "utf-8")
-
-            parser.set_config(config)
-
-            # Test parsing with config
-            result = parser.parse("# Test")
-            assert result is not None
-
-        except AttributeError:
-            # Config integration might not be implemented
-            pass
-
-    def test_config_with_renderer(self):
-        """Test config integration with renderer"""
-        from kumihan_formatter.config import ConfigManager
-        from kumihan_formatter.core.ast_nodes.node import Node
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-        config = ConfigManager()
-
-        # Test renderer accepts config
-        try:
-            config.set("template", "default")
-            config.set("minify", False)
-
-            renderer.set_config(config)
-
-            # Test rendering with config
-            nodes = [Node("p", "Test")]
-            result = renderer.render(nodes)
-            assert isinstance(result, str)
-
-        except AttributeError:
-            # Config integration might not be implemented
-            pass
-
-    def test_config_file_loading(self):
-        """Test loading config from file"""
-        from kumihan_formatter.config.config_manager import ConfigManager
-
-        config_manager = ConfigManager()
-
-        # Create temporary config file
-        config_data = {
-            "output_format": "html",
-            "encoding": "utf-8",
-            "template": "default",
-        }
-
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".json", delete=False) as f:
-            import json
-
-            json.dump(config_data, f)
-            temp_path = f.name
-
-        try:
-            # Test loading from file
-            config_manager.load_from_file(temp_path)
-
-            # Verify loaded data
-            assert config_manager.get("output_format") == "html"
-            assert config_manager.get("encoding") == "utf-8"
-
-        except AttributeError:
-            # File loading might not be implemented
-            pass
-        finally:
-            Path(temp_path).unlink(missing_ok=True)
-
-    def test_config_environment_variables(self):
-        """Test config from environment variables"""
-        import os
-
-        from kumihan_formatter.config.config_manager_env import ConfigManagerEnv
-
-        try:
-            config_manager = ConfigManagerEnv()
-
-            # Set test environment variables
-            os.environ["KUMIHAN_OUTPUT_FORMAT"] = "html"
-            os.environ["KUMIHAN_ENCODING"] = "utf-8"
-
-            try:
-                config_manager.load_from_env("KUMIHAN_")
-
-                # Verify environment config
-                output_format = config_manager.get("output_format")
-                encoding = config_manager.get("encoding")
-
-                assert output_format == "html"
-                assert encoding == "utf-8"
-
-            except AttributeError:
-                pass
-            finally:
-                # Cleanup environment variables
-                os.environ.pop("KUMIHAN_OUTPUT_FORMAT", None)
-                os.environ.pop("KUMIHAN_ENCODING", None)
-
-        except ImportError:
-            # Module might not exist
-            pass
