--- tests/test_realistic_renderer_coverage.py
@@ -1,337 +1,0 @@
-"""Realistic Renderer Coverage Tests
-
-Focus on actual working renderer functionality with realistic test cases.
-Target: Increase renderer module coverage significantly.
-"""
-
-import pytest
-
-from kumihan_formatter.core.ast_nodes.node import Node
-
-
-class TestRendererRealUsage:
-    """Test renderer with real-world usage patterns"""
-
-    def test_renderer_basic_functionality(self):
-        """Test basic renderer functionality"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test simple node rendering
-        simple_nodes = [
-            [Node("p", "Simple paragraph")],
-            [Node("h1", "Main Heading")],
-            [Node("h2", "Sub Heading")],
-            [Node("div", "Container content")],
-            [Node("span", "Inline content")],
-        ]
-
-        for nodes in simple_nodes:
-            result = renderer.render(nodes)
-            assert isinstance(result, str)
-            assert len(result) > 0
-
-            # Should contain the content
-            for node in nodes:
-                if isinstance(node.content, str):
-                    assert node.content in result or len(node.content) == 0
-
-    def test_renderer_heading_rendering(self):
-        """Test heading rendering"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test different heading levels
-        for level in range(1, 6):
-            nodes = [Node(f"h{level}", f"Heading Level {level}")]
-            result = renderer.render(nodes)
-
-            assert isinstance(result, str)
-            assert f"Heading Level {level}" in result
-
-    def test_renderer_list_rendering(self):
-        """Test list rendering"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test unordered list
-        ul_nodes = [
-            Node(
-                "ul",
-                [
-                    Node("li", "Item 1"),
-                    Node("li", "Item 2"),
-                    Node("li", "Item 3"),
-                ],
-            )
-        ]
-
-        result = renderer.render(ul_nodes)
-        assert isinstance(result, str)
-        assert "Item 1" in result
-        assert "Item 2" in result
-        assert "Item 3" in result
-
-        # Test ordered list
-        ol_nodes = [
-            Node(
-                "ol",
-                [
-                    Node("li", "First"),
-                    Node("li", "Second"),
-                    Node("li", "Third"),
-                ],
-            )
-        ]
-
-        result = renderer.render(ol_nodes)
-        assert isinstance(result, str)
-        assert "First" in result
-        assert "Second" in result
-        assert "Third" in result
-
-    def test_renderer_nested_content(self):
-        """Test nested content rendering"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test nested structure
-        nested_nodes = [
-            Node(
-                "div",
-                [
-                    Node("h2", "Section Title"),
-                    Node("p", "Section content goes here."),
-                    Node(
-                        "div",
-                        [
-                            Node("h3", "Subsection"),
-                            Node("p", "Subsection content."),
-                        ],
-                    ),
-                ],
-            )
-        ]
-
-        result = renderer.render(nested_nodes)
-        assert isinstance(result, str)
-        assert "Section Title" in result
-        assert "Section content goes here." in result
-        assert "Subsection" in result
-        assert "Subsection content." in result
-
-    def test_renderer_attributes(self):
-        """Test rendering nodes with attributes"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Create nodes with attributes
-        div_node = Node("div", "Content with attributes")
-        div_node.add_attribute("class", "container")
-        div_node.add_attribute("id", "main-content")
-
-        link_node = Node("a", "Link text")
-        link_node.add_attribute("href", "https://example.com")
-        link_node.add_attribute("target", "_blank")
-
-        nodes = [div_node, link_node]
-
-        result = renderer.render(nodes)
-        assert isinstance(result, str)
-        assert "Content with attributes" in result
-        assert "Link text" in result
-
-    def test_renderer_complex_document(self):
-        """Test rendering complex document"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Create realistic document structure
-        document_nodes = [
-            Node("h1", "Document Title"),
-            Node("p", "This is the introduction paragraph with important information."),
-            Node("h2", "Features"),
-            Node(
-                "ul",
-                [
-                    Node("li", "Feature 1: Basic functionality"),
-                    Node("li", "Feature 2: Advanced options"),
-                    Node("li", "Feature 3: Integration support"),
-                ],
-            ),
-            Node("h2", "Usage"),
-            Node("p", "Here's how to use the features:"),
-            Node(
-                "ol",
-                [
-                    Node("li", "Step 1: Initialize the system"),
-                    Node("li", "Step 2: Configure options"),
-                    Node("li", "Step 3: Execute operations"),
-                ],
-            ),
-            Node(
-                "div",
-                [
-                    Node("h3", "Important Note"),
-                    Node("p", "Please read the documentation carefully."),
-                ],
-            ),
-        ]
-
-        result = renderer.render(document_nodes)
-        assert isinstance(result, str)
-        assert len(result) > 100  # Should be substantial content
-
-        # Verify all content is present
-        content_checks = [
-            "Document Title",
-            "introduction paragraph",
-            "Features",
-            "Feature 1",
-            "Feature 2",
-            "Feature 3",
-            "Usage",
-            "Step 1",
-            "Step 2",
-            "Step 3",
-            "Important Note",
-            "documentation carefully",
-        ]
-
-        for content in content_checks:
-            assert content in result
-
-
-class TestHTMLRenderer:
-    """Test HTML renderer specifically"""
-
-    def test_html_renderer_basic(self):
-        """Test basic HTML renderer functionality"""
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        renderer = HTMLRenderer()
-
-        # Test single node rendering
-        node = Node("p", "Test paragraph")
-
-        try:
-            result = renderer.render_node(node)
-            assert isinstance(result, str)
-            assert "Test paragraph" in result
-        except AttributeError:
-            # Method might be named differently
-            try:
-                result = renderer.render([node])
-                assert isinstance(result, str)
-            except:
-                pass
-
-    def test_html_renderer_nodes_list(self):
-        """Test HTML renderer with nodes list"""
-        from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-
-        renderer = HTMLRenderer()
-
-        nodes = [
-            Node("h1", "Title"),
-            Node("p", "Paragraph 1"),
-            Node("p", "Paragraph 2"),
-        ]
-
-        try:
-            result = renderer.render_nodes(nodes)
-            assert isinstance(result, str)
-            assert "Title" in result
-            assert "Paragraph 1" in result
-            assert "Paragraph 2" in result
-        except AttributeError:
-            # Method might not exist
-            pass
-
-
-class TestRendererIntegration:
-    """Test renderer integration with other components"""
-
-    def test_render_function_integration(self):
-        """Test the render() function integration"""
-        from kumihan_formatter.renderer import Renderer
-
-        # Test basic functionality
-        nodes = [
-            Node("h1", "Test Document"),
-            Node("p", "Test content"),
-        ]
-
-        renderer = Renderer()
-        result = renderer.render(nodes)
-        assert isinstance(result, str)
-        assert "Test Document" in result
-        assert "Test content" in result
-
-    def test_renderer_with_config(self):
-        """Test renderer with configuration"""
-        from kumihan_formatter.config import ConfigManager
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-        config = ConfigManager()
-
-        # Test renderer accepts config (if supported)
-        try:
-            renderer.set_config(config)
-        except AttributeError:
-            # Config might not be supported
-            pass
-
-        # Render with potential config
-        nodes = [Node("p", "Test with config")]
-        result = renderer.render(nodes)
-        assert isinstance(result, str)
-
-    def test_renderer_template_system(self):
-        """Test renderer template system"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-        nodes = [Node("p", "Template test")]
-
-        # Try different templates
-        templates = ["default", "minimal", "simple"]
-
-        for template in templates:
-            try:
-                renderer.set_template(template)
-                result = renderer.render(nodes)
-                assert isinstance(result, str)
-            except AttributeError:
-                # Template system might not be implemented
-                pass
-
-    def test_renderer_error_handling(self):
-        """Test renderer error handling"""
-        from kumihan_formatter.renderer import Renderer
-
-        renderer = Renderer()
-
-        # Test edge cases
-        edge_cases = [
-            [],  # Empty nodes list
-            [Node("", "")],  # Empty node type
-            [Node("unknown", "content")],  # Unknown node type
-            [Node("p", None)],  # None content
-            [Node("div", [])],  # Empty children
-        ]
-
-        for nodes in edge_cases:
-            try:
-                result = renderer.render(nodes)
-                assert isinstance(result, str)
-            except (TypeError, AttributeError):
-                # Some edge cases might not be handled
-                pass
