+++ tests/test_specific_module_coverage.py
@@ -1,193 +1,12 @@
-"""Parser and Specific Module Integration Tests
+"""Specific Module Coverage Tests
 
-Deep integration tests for parser systems and specific modules.
-Targets specific uncovered code paths for maximum coverage impact.
+Target specific modules for coverage improvement.
+Tests for TOC generators and list parsers.
 """
 
 from unittest.mock import Mock, patch
 
 import pytest
-
-
-class TestParserComprehensiveIntegration:
-    """Comprehensive parser testing for coverage boost"""
-
-    def test_kumihan_parser_main_functionality(self):
-        """Test main Kumihan parser functionality"""
-        try:
-            from kumihan_formatter.parser import Parser
-        except ImportError as e:
-            pytest.skip(f"Parser not available: {e}")
-            return
-
-        parser = Parser()
-
-        # Test various Kumihan syntax patterns
-        test_inputs = [
-            "Simple paragraph text",
-            "# Heading level 1",
-            "## Heading level 2",
-            "### Heading level 3",
-            "**Bold text**",
-            "*Italic text*",
-            ";;;highlight;;; highlighted content ;;;",
-            ";;;box;;; boxed content ;;;",
-            ";;;footnote;;; footnote text ;;;",
-            "((footnote content))",
-            "｜ruby《reading》",
-            "Multiple\nlines\nof\ncontent",
-            """Multi-paragraph content
-
-With blank lines
-
-And different formatting elements.""",
-        ]
-
-        for test_input in test_inputs:
-            try:
-                result = parser.parse(test_input)
-                assert result is not None
-
-                # Result should be a list of nodes or similar structure
-                if hasattr(result, "__len__"):
-                    assert len(result) >= 0
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                # Some parsing may not be fully implemented
-                pytest.skip(f"Parser test failed for input: {test_input[:30]}... - {e}")
-
-        # Test parser configuration
-        try:
-            parser.set_config({"strict_mode": True})
-            parser.set_config({"allow_html": False})
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-    def test_renderer_main_functionality(self):
-        """Test main renderer functionality"""
-        try:
-            from kumihan_formatter.core.ast_nodes.node import Node
-            from kumihan_formatter.renderer import Renderer
-        except ImportError as e:
-            pytest.skip(f"Renderer not available: {e}")
-            return
-
-        renderer = Renderer()
-
-        # Test rendering different node types
-        test_nodes = [
-            [Node("p", "Simple paragraph")],
-            [Node("h1", "Main heading")],
-            [Node("h2", "Sub heading")],
-            [Node("div", "Container content")],
-            [Node("strong", "Bold text")],
-            [Node("em", "Italic text")],
-            [
-                Node(
-                    "div", [Node("h1", "Nested heading"), Node("p", "Nested paragraph")]
-                )
-            ],
-            [Node("ul", [Node("li", "List item 1"), Node("li", "List item 2")])],
-        ]
-
-        for nodes in test_nodes:
-            try:
-                result = renderer.render(nodes)
-                assert isinstance(result, str)
-                assert len(result) > 0
-
-                # Should contain some HTML-like structure
-                if nodes[0].type in ["p", "h1", "h2", "div"]:
-                    assert f"<{nodes[0].type}>" in result or nodes[0].type in result
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                pytest.skip(
-                    f"Renderer test failed for node type: {nodes[0].type} - {e}"
-                )
-
-        # Test renderer configuration
-        try:
-            renderer.set_template("minimal")
-            renderer.set_template("default")
-        except (
-            AttributeError,
-            NotImplementedError,
-            TypeError,
-            ValueError,
-            ImportError,
-        ) as e:
-            pass
-
-    def test_parse_render_workflow_comprehensive(self):
-        """Test complete parse-render workflow"""
-        try:
-            from kumihan_formatter.parser import Parser
-            from kumihan_formatter.renderer import Renderer
-        except ImportError as e:
-            pytest.skip(f"Parser/Renderer not available: {e}")
-            return
-
-        parser = Parser()
-        renderer = Renderer()
-
-        # Test complete workflow with various inputs
-        workflow_tests = [
-            "# Document Title\n\nThis is a paragraph with **bold** text.",
-            ";;;highlight;;; Important information ;;;",
-            "Multiple paragraphs\n\nWith different content\n\nAnd formatting.",
-            "## Section\n\n- List item 1\n- List item 2",
-            "((This is a footnote))",
-            "｜日本語《にほんご》のルビ表記",
-        ]
-
-        for test_input in workflow_tests:
-            try:
-                # Parse input
-                parsed_nodes = parser.parse(test_input)
-                assert parsed_nodes is not None
-
-                # Render parsed nodes
-                rendered_output = renderer.render(parsed_nodes)
-                assert isinstance(rendered_output, str)
-                assert len(rendered_output) > 0
-
-                # Basic sanity checks
-                if "**" in test_input:
-                    # Bold text should be processed somehow
-                    assert "**" not in rendered_output or "strong" in rendered_output
-
-                if ";;;" in test_input:
-                    # Kumihan syntax should be processed
-                    assert (
-                        ";;;" not in rendered_output or "highlight" in rendered_output
-                    )
-
-            except (
-                AttributeError,
-                NotImplementedError,
-                TypeError,
-                ValueError,
-                ImportError,
-            ) as e:
-                pytest.skip(f"Workflow test failed for: {test_input[:30]}... - {e}")
 
 
 class TestSpecificModuleCoverage:
